Just a small little commit.
Or, alternatively, what happens when I don't commit before moving on
to the next task.

Make it safe to use gcc 3.1! (Even with -march=athlon!!)

Steal the collision code from DP and use it for both NQ and QW.
(With several modifications.)

Killed several functions from mathlib which did nothing but call the C
counterparts, this is a speed gain.

vec_t is now based on the GL types.
dvec_t and friends for double vectors, bvec_t for unsigned byte vectors.

Union type float_int_t for when we have something treated as both an int
and a float. (For instance, Q_RSqrt) (Needed for gcc 3.1)

Replace some users of malloc with Zone_Alloc.

We now use unsigned bytes for vertex arrays instead of floats.
(We also bound them properly, which should kill twibugs_evilflames)

Don't draw stainmaps either if they are disabled.

Some diff reduction between QW and NQ.

Optimize particles to properly die, speed up and bug fix in one!

Some sbar optimizations.

In several places in qw and qw-server now use entity pointers instead of
indices into an array of entities, for the collision code stuff.

Probably a few other things too, but this is the big stuff.


git-svn-id: svn://svn.icculus.org/twilight/trunk@1896 d7cf8633-e32d-0410-b094-e92efae38249