Oh my god, they killed qw-server!  (YOU BASTARDS!)
Yes, that's right, qw-server is dead.  qw now builds both a client and a
server.  The client does not have an internal server yet, but the
framework for doing that is laid in this commit.  Cvar con_logname has
been renamed to sys_logname, but its usage is unchanged.  include/sys.h
now includes a few defines and a prototype for the type of game that's
currently running, but NQ doesn't use it yet.


git-svn-id: svn://svn.icculus.org/twilight/trunk@1921 d7cf8633-e32d-0410-b094-e92efae38249