Big revert.  Here lies lesson one: Fuck not with that ye dost not fully understand.  I tried to preserve changes which did not break things, but several useful things (signed int to unsigned int conversions) were lost. That would make lesson two: Commit thy cleanups seperately from thy changes if thou wishest not to lose these when thy bugs are reverted.
I've spent several hours reverting this mess by hand and I'm not happy
about it.  That makes lesson three: Maketh not thy commits too large lest
Knghtbrd need to kill thee.  As much as Zeph will not want to hear it, I'm
going to suggest that the interesting parts of his changes be reapplied
carefully.  As for the rest, I think we have found his implementation to
not work right.  I suggest we start over on this front, AFTER 0.2 is out.


git-svn-id: svn://svn.icculus.org/twilight/trunk@2122 d7cf8633-e32d-0410-b094-e92efae38249