There is a brief bright flash, and then nothing.
This is the sound of a 50-kiloton explosion at medium range.

There is no sound because the blast wave travels faster then the
speed of sound.

The device was attached directly to the hunk and cache code.
Sadly, nothing survived the blast.

I apologize for Zone_AllocName, and the name of COM_LoadNamedFile.

This also fixes a bug with maps, say you are on map a, switch to map b,
then switch back to map a, you had a severe problem before, fixed now.


git-svn-id: svn://svn.icculus.org/twilight/trunk@2222 d7cf8633-e32d-0410-b094-e92efae38249