Fix RC #0!
The problem was actually with OP_STORE_V, this copys a vector, and is used by qcc during function calls.
The problem is that for some reason gcc 2.95 generates code that slightly mangles 'invalid' floats for this copy, and the result of some of the pointer calcs was decided 'invalid' as far as a float goes.
So, we now copy as ints, and the problem is GONE!

(This resulted in making PR_GetString and PR_SetString #defines for QW and DP, making for cleaner code and less memory usage.)


git-svn-id: svn://svn.icculus.org/twilight/trunk@2705 d7cf8633-e32d-0410-b094-e92efae38249