This is sadly /TWO/ huge changes at once.
The first is the scrapping of the old v_gamma and v_tgamma.
Replacing it are a few settings:
black, grey, and while.  With v_ and v_t prefixes, and blank, _r, _g,
and _b suffixes.


The second is the complete scrapping of the old menu system.

Replacing it is a /much/ easier to change system, the parsing is a
major PITA, however the rendering and changing of the menu itself is
/very/ simple.

The command to load a menu from a text file is 'loadmenu <file>'.

Included is an example menu file, menu.mnu, shove it in your id1/ and
load it to get a basic menu for both QW and NQ.

There is still a todo list:
 - Get the renderer side of the menu code to handle menus larger then
   the screen.
 - Find a clean way to include a text file in the binary so that we
   still have a basic menu.
 - Add a command to /delete/ a menu. (Mostly trivial.)
 - Saved games support.
 - Server browser support.
 - Key binding menu support.


git-svn-id: svn://svn.icculus.org/twilight/trunk@2849 d7cf8633-e32d-0410-b094-e92efae38249