The main bulk of my changes.
Don't blank r_skyname on map changes, if someone sets things to use a
skybox, then don't force it to skyspheres.

Make everything /except/ lightmaps use things like GLT_Load_* and
GL_Upload32 instead of uploading directly, this makes sure that we
honor the gl_max_size on everything except lightmaps.

Make sure that gl_max_size is a power of 2.

Move some variable declarations.

Change how we handle CVAs a bit, this should work with nVidia drivers
and be faster in a number of cases. (Though, for some reason I'm not
seeing much in the way of speedups.)

Change some of the vertex array handling.

Added TEX_REPLACE flag, this causes the texture to replace other
textures with the same name, used for skys.

Change some of the sky sphere drawing, a little bit faster.


git-svn-id: svn://svn.icculus.org/twilight/trunk@2956 d7cf8633-e32d-0410-b094-e92efae38249