Always unload the worldmodel (and submodels) when we clear the variable.
Don't check to see if mdl_torch is non-NULL when we load flame.mdl, as it gets unloaded when we do a Mod_ClearAll.

Make mod_known and mod_knownall static.

Some model loading/unloading debugging stuff for nq.

Fix a qw bug dealing with the entity list and U_REMOVE on full update that sometimes caused interesting crashes.


git-svn-id: svn://svn.icculus.org/twilight/trunk@5533 d7cf8633-e32d-0410-b094-e92efae38249