For some reason, cl_netinputpacketsperserverpacket > 1 causes choppiness when cl_movement is active and fps are high.
As workaround, disable the "replying to server frames" code for cl_netinputpacketsperserverpacket > 1, and make it ONLY respond to server frames (or after 0.1s when no server frame came) when set to 0. 0 MAY be a bit more smooth than 1, but something tells me it is LESS smooth when packet loss is involved.


git-svn-id: svn://svn.icculus.org/twilight/trunk@7985 d7cf8633-e32d-0410-b094-e92efae38249