first test of multi-account rcon:
	rcon_password "account1:1234 account2:2345"

Any of the space separated passwords can be matched - the string before the :
in each 'password' is printed to the log. So if that is set, clients can send
commands both with rcon_password "account1:1234" as well as with rcon_password
"account2:2345".

Fully backwards compatible: if rcon_password contains no space, user name
printing is not performed to keep the logs clean.

Also supported for rcon_restricted of course.

NOTE: anyone with rcon_password can read the passwords of the others by simply
sending the command "rcon_password". Of course this will get logged, but makes
all other passwords pointless. Better use this system only for RESTRICTED rcon,
is what I am saying. And WATCH YOUR LOGS.



git-svn-id: svn://svn.icculus.org/twilight/trunk@9420 d7cf8633-e32d-0410-b094-e92efae38249