New alternative to fullbright: r_fullbright_directed.
Set this to 1, and anything EF_FULLBRIGHT or fullbright due to
r_fullbright or unlit map (but specifically not sprites and surfaces
with surfaceparm nolightmap) will have a directional effect applied to
them.

This mode is intended for testing maps in early development, and should
not be used for normal gameplay (although it should be rather harmless
there).

git-svn-id: svn://svn.icculus.org/twilight/trunk@12364 d7cf8633-e32d-0410-b094-e92efae38249